test: Move batcher fork safety test to batcher tests
